requirements/base.txt
dartcms
django-rosetta
django-subdomains
psycopg2